+++ gym_donkeycar/envs/donkey_sim.py
@@ -127,6 +127,12 @@ class DonkeyUnitySimHandler(IMesgHandler):
         self.vel_x = 0.0
         self.vel_y = 0.0
         self.vel_z = 0.0
+        self.lidar = []
+
+        # car in Unity lefthand coordinate system: roll is Z, pitch is X and yaw is Y
+        self.roll = 0.0
+        self.pitch = 0.0
+        self.yaw = 0.0
 
     def on_connect(self, client: SimClient) -> None:
         logger.debug("socket connected")
@@ -172,7 +178,48 @@ class DonkeyUnitySimHandler(IMesgHandler):
                 "rot_x",
             ],
         )
-        self.send_cam_config(**cam_config)
+        try:
+            cam_config = self.extract_keys(
+                conf,
+                [
+                    "img_w",
+                    "img_h",
+                    "img_d",
+                    "img_enc",
+                    "fov",
+                    "fish_eye_x",
+                    "fish_eye_y",
+                    "offset_x",
+                    "offset_y",
+                    "offset_z",
+                    "rot_x",
+                ],
+            )
+            self.send_cam_config(**cam_config)
+            logger.info("done sending cam config.", cam_config)
+        except:
+            logger.info("sending cam config FAILED.")
+
+        try:
+            lidar_config = self.extract_keys(
+                conf,
+                [
+                    "deg_per_sweep_inc",
+                    "deg_ang_down",
+                    "deg_ang_delta",
+                    "num_sweeps_levels",
+                    "max_range",
+                    "noise",
+                    "offset_x",
+                    "offset_y",
+                    "offset_z",
+                    "rot_x",
+                ],
+            )
+            self.send_lidar_config(**lidar_config)
+            logger.info("done sending lidar config.", lidar_config)
+        except:
+            logger.info("sending lidar config FAILED.")
         logger.info("done sending car config.")
 
     def set_car_config(self, conf: Dict[str, Any]) -> None:
@@ -221,6 +268,12 @@ class DonkeyUnitySimHandler(IMesgHandler):
         self.vel_x = 0.0
         self.vel_y = 0.0
         self.vel_z = 0.0
+        self.lidar = []
+
+        # car
+        self.roll = 0.0
+        self.pitch = 0.0
+        self.yaw = 0.0
 
     def get_sensor_size(self) -> Tuple[int, int, int]:
         return self.camera_img_size
@@ -246,6 +299,8 @@ class DonkeyUnitySimHandler(IMesgHandler):
             "gyro": (self.gyro_x, self.gyro_y, self.gyro_z),
             "accel": (self.accel_x, self.accel_y, self.accel_z),
             "vel": (self.vel_x, self.vel_y, self.vel_z),
+            "lidar": (self.lidar),
+            "car": (self.roll, self.pitch, self.yaw),
         }
 
         # self.timer.on_frame()
@@ -305,11 +360,19 @@ class DonkeyUnitySimHandler(IMesgHandler):
             self.vel_y = message["vel_y"]
             self.vel_z = message["vel_z"]
 
+        if "roll" in data:
+            self.roll = data["roll"]
+            self.pitch = data["pitch"]
+            self.yaw = data["yaw"]
+
         # Cross track error not always present.
         # Will be missing if path is not setup in the given scene.
         # It should be setup in the 4 scenes available now.
         if "cte" in message:
             self.cte = message["cte"]
+
+        if "lidar" in data:
+            self.lidar = data["lidar"]
 
         # don't update hit once session over
         if self.over:
@@ -474,6 +537,57 @@ class DonkeyUnitySimHandler(IMesgHandler):
             "img_h": str(img_h),
             "img_d": str(img_d),
             "img_enc": str(img_enc),
+            "offset_x": str(offset_x),
+            "offset_y": str(offset_y),
+            "offset_z": str(offset_z),
+            "rot_x": str(rot_x),
+        }
+        self.blocking_send(msg)
+        time.sleep(0.1)
+
+    def send_lidar_config(
+        self,
+        deg_per_sweep_inc,
+        deg_ang_down,
+        deg_ang_delta,
+        num_sweeps_levels,
+        max_range,
+        noise,
+        offset_x,
+        offset_y,
+        offset_z,
+        rot_x,
+    ):
+        """Lidar config
+        offset_x moves lidar left/right
+        the offset_y moves lidar up/down
+        the offset_z moves lidar forward/back
+        deg_per_sweep_inc : as the ray sweeps around, how many degrees does it advance per sample (int)
+        deg_ang_down : what is the starting angle for the initial sweep compared to the forward vector
+        deg_ang_delta : what angle change between sweeps
+        num_sweeps_levels : how many complete 360 sweeps (int)
+        max_range : what it max distance we will register a hit
+        noise : what is the scalar on the perlin noise applied to point position
+
+        Here's some sample settings that similate a more sophisticated lidar:
+        msg = '{ "msg_type" : "lidar_config",
+        "degPerSweepInc" : "2.0", "degAngDown" : "25", "degAngDelta" : "-1.0",
+        "numSweepsLevels" : "25", "maxRange" : "50.0", "noise" : "0.2",
+        "offset_x" : "0.0", "offset_y" : "1.0", "offset_z" : "1.0", "rot_x" : "0.0" }'
+        And here's some sample settings that similate a simple RpLidar A2 one level horizontal scan.
+        msg = '{ "msg_type" : "lidar_config", "degPerSweepInc" : "2.0",
+        "degAngDown" : "0.0", "degAngDelta" : "-1.0", "numSweepsLevels" : "1",
+        "maxRange" : "50.0", "noise" : "0.4",
+        "offset_x" : "0.0", "offset_y" : "0.5", "offset_z" : "0.5", "rot_x" : "0.0" }'
+        """
+        msg = {
+            "msg_type": "lidar_config",
+            "degPerSweepInc": str(deg_per_sweep_inc),
+            "degAngDown": str(deg_ang_down),
+            "degAngDelta": str(deg_ang_delta),
+            "numSweepsLevels": str(num_sweeps_levels),
+            "maxRange": str(max_range),
+            "noise": str(noise),
             "offset_x": str(offset_x),
             "offset_y": str(offset_y),
             "offset_z": str(offset_z),